backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.3.1
    # via
    #   keras
    #   tensorboard
    #   tensorflow-intel
astunparse==1.6.3
    # via tensorflow-intel
beautifulsoup4==4.13.4
    # via gdown
blinker==1.9.0
    # via flask
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via flask
colorama==0.4.6
    # via
    #   click
    #   tqdm
deepface==0.0.87
    # via -r requirements.in
filelock==3.18.0
    # via gdown
fire==0.7.0
    # via deepface
flask==2.3.3
    # via
    #   -r requirements.in
    #   deepface
    #   flask-cors
flask-cors==4.0.0
    # via -r requirements.in
flatbuffers==25.2.10
    # via tensorflow-intel
gast==0.6.0
    # via tensorflow-intel
gdown==4.7.1
    # via
    #   -r requirements.in
    #   deepface
    #   retina-face
google-pasta==0.2.0
    # via tensorflow-intel
grpcio==1.73.1
    # via
    #   tensorboard
    #   tensorflow-intel
gunicorn==23.0.0
    # via deepface
h5py==3.14.0
    # via
    #   keras
    #   tensorflow-intel
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
keras==3.10.0
    # via
    #   deepface
    #   mtcnn
    #   tensorflow-intel
libclang==18.1.1
    # via tensorflow-intel
markdown==3.8.2
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.3.2
    # via
    #   keras
    #   tensorflow-intel
mtcnn==0.1.1
    # via deepface
namex==0.1.0
    # via keras
numpy==1.26.4
    # via
    #   -r requirements.in
    #   deepface
    #   h5py
    #   keras
    #   ml-dtypes
    #   opencv-python
    #   opencv-python-headless
    #   pandas
    #   retina-face
    #   tensorboard
    #   tensorflow-intel
opencv-python==4.11.0.86
    # via
    #   deepface
    #   mtcnn
    #   retina-face
opencv-python-headless==4.9.0.80
    # via -r requirements.in
opt-einsum==3.4.0
    # via tensorflow-intel
optree==0.16.0
    # via keras
packaging==25.0
    # via
    #   gunicorn
    #   keras
    #   tensorflow-intel
pandas==2.3.1
    # via deepface
pillow==11.3.0
    # via
    #   deepface
    #   retina-face
protobuf==4.25.3
    # via
    #   -r requirements.in
    #   tensorboard
    #   tensorflow-intel
pydantic==1.10.13
    # via -r requirements.in
pygments==2.19.2
    # via rich
pysocks==1.7.1
    # via requests
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
requests[socks]==2.32.4
    # via
    #   gdown
    #   tensorflow-intel
retina-face==0.0.17
    # via deepface
rich==14.0.0
    # via keras
six==1.17.0
    # via
    #   astunparse
    #   gdown
    #   google-pasta
    #   python-dateutil
    #   tensorboard
    #   tensorflow-intel
soupsieve==2.7
    # via beautifulsoup4
tensorboard==2.16.2
    # via tensorflow-intel
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.16.1
    # via
    #   -r requirements.in
    #   deepface
    #   retina-face
tensorflow-intel==2.16.1
    # via tensorflow
termcolor==3.1.0
    # via
    #   fire
    #   tensorflow-intel
tqdm==4.67.1
    # via
    #   deepface
    #   gdown
typing-extensions==4.14.1
    # via
    #   beautifulsoup4
    #   optree
    #   pydantic
    #   tensorflow-intel
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via
    #   flask
    #   tensorboard
wheel==0.45.1
    # via astunparse
wrapt==1.17.2
    # via tensorflow-intel

tf-keras
ultralytics opencv-python-headless numpy
face-alignment
openai>=1.0.0

# The following packages are considered to be unsafe in a requirements file:
# setuptools
